Option appendToListenerForwarder for ProgressListenerConfiguration.
When you add the **appendToListenerForwarder** flag for your listener configuration, the listener it's added to the listenerForwarder too. This option allows manage the event status and messages from the custom actions.

To enable it, you need to add listeners at the buttom of the flow declaration and reference it from your action. This example show how to add in a flow configuration:
´´´
<FlowConfiguration>
        ...
	<EventConsumerConfiguration>
		<YourActionConfiguration>
			...
			<!-- Listeners -->
			<listenerId>Cumulator</listenerId>
		</YourActionConfiguration>
	</EventConsumerConfiguration>
	<ListenerConfigurations>
		<CumulatingProgressListener>
			<serviceID>cumulatingListenerService</serviceID>
			<id>Cumulator</id>
			<appendToListenerForwarder>true</appendToListenerForwarder>
		</CumulatingProgressListener>
	</ListenerConfigurations>
</FlowConfiguration>
´´´